modules/iam/templates/requirements.txt
google-cloud-resource-manager==1.10.4
google-cloud-iam==2.12.1
google-cloud-bigquery==3.11.4
google-cloud-logging==3.8.0
functions-framework==3.4.0